Add copyright notices regarding the words file
They are all FOSS permissive licenses
(and appear GPL-compatible, although the GPL's copyleft does
not travel between code and the data files the code uses,
so even if my code were GPL it wouldn't matter.).

It never hurts to put in copyright notices even if they're
not necessary.  The words file I use currently is
derived from the one covered by this copyright notice,
with an attempt by me to remove potentially offensive or
disturbing words; my purpose for the list is not at all to be
a comprehensive list of words, but rather to be a list of words
I could reasonably randomly generate for user things.